Changed int8 to uint in contracts and models

+++ contracts/training/create_training.go
@@ -11,7 +11,7 @@ type CreateTrainingRequest struct {
 	Description string            `json:"description" binding:"required"`
 	Difficulty  string            `json:"difficulty" binding:"required"`
 	TrainerID   string            `json:"trainer_id" binding:"required"`
-	Duration    int8              `json:"duration" binding:"required"`
+	Duration    uint              `json:"duration" binding:"required"`
 	Exercises   []ExerciseRequest `json:"exercises" binding:"required"`
 }
